# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Inventory Management API",
        "description": "Documentació oficial de l'API d'Inventari - Projecte Final",
        "contact": {
            "email": "admin@admin.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/owners": {
            "get": {
                "tags": [
                    "Owners"
                ],
                "summary": "Obtenir llista de propietaris",
                "operationId": "44d28782c206624819269395b7f4cd82",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Llista de propietaris",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Owner"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Owners"
                ],
                "summary": "Crear un nou propietari",
                "operationId": "c7eb4ec9f7e6a1b9d7fbe0090cfbb7da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john@doe.com"
                                    },
                                    "department": {
                                        "type": "string",
                                        "example": "Sales"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creat correctament"
                    },
                    "422": {
                        "description": "Error de validació"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/owners/{id}": {
            "get": {
                "tags": [
                    "Owners"
                ],
                "summary": "Detalls d'un propietari",
                "operationId": "0ce511882e5365c1c4e09c1be6e59c4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "No trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Owners"
                ],
                "summary": "Actualitzar propietari",
                "operationId": "4b0b6bc791a9fa75f625c9827511f7b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "department": {
                                        "type": "string",
                                        "example": "Human Resources"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualitzat correctament"
                    },
                    "404": {
                        "description": "No trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Owners"
                ],
                "summary": "Esborrar propietari",
                "operationId": "e1cec6afe0543d87ab57e37b41f2b716",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Esborrat correctament"
                    },
                    "404": {
                        "description": "No trobat"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Owner": {
                "title": "Owner",
                "description": "Owner model definition",
                "required": [
                    "owner_name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "owner_name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "department": {
                        "type": "string",
                        "example": "IT"
                    },
                    "location": {
                        "type": "string",
                        "example": "Barcelona"
                    },
                    "employee_code": {
                        "type": "string",
                        "example": "EMP-001"
                    },
                    "phone": {
                        "type": "string",
                        "example": "600123456"
                    },
                    "extension": {
                        "type": "string",
                        "example": "101"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "notes": {
                        "type": "string",
                        "example": "Contractat el 2023"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}